examples/custom_expression/pyproject.toml
[tool.poetry]
name = "polars_casing"
version = "0.1.0"
description = ""
authors = ["Jan Vanbuel <janvanbuel@hotmail.com>"]
readme = "README.md"
packages = [{ include = "polars_casing" }]

[tool.poetry.dependencies]
python = "^3.10"
maturin = "^1.3"
polars = "^0.19.7"
cffi = "^1.16.0"
wonderwords = "^2.2.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.27.1"
pytest-benchmark = "^4.0.0"
isort = "^5.12.0"

[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

